side/requirements.txt
gradio==3.27.0
PyPDF2==3.0.1
pymysql==1.0.3
re
difflib
langchain_community
ollama
